feat: complete order flow with validations and negative cases

+++ utown_backend_nov18/src/main/java/org/example/utown_backend_nov18/dto/OrderItemDto.java
@@ -1,6 +1,7 @@
 package org.example.utown_backend_nov18.dto;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class OrderItemDto {
@@ -8,5 +9,5 @@ public class OrderItemDto {
     private Long dishId;
     private String dishName;
     private Integer quantity;
-    private Integer priceAtMoment;
+    private BigDecimal priceAtMoment;
 }
